feat(core-elements): export NumericSpinnerBase
