+++ packages/core-elements/src/elements/numeric-spinner/index.ts
@@ -1,1 +1,2 @@
+export * from './numeric-spinner-base'
 export * from './numeric-spinner'
